Added a new bibconver formatting function "IFDEFP". It is like a mixture of DEFP (which signals that a line should always be printed) and CONF (which prints a line based upon whether a condition met by the contents of another field variable is met). So, this function IFDEFP is used on constant lines to say that they should be printed based upon the contents of another field. E.g. given the following line: 690C::REP(EOL,)::IFDEFP(comboYEL,BOOK,1)---<datafield tag="690" ind1="C" ind2=" "><subfield code="a">BOOK</subfield></datafield>
We want to print the line if the "comboYEL" field contains "BOOK", otherwise
we don't want to print it.  This is achieved by using "IFDEFP". If the line
had contained variables, "CONF" would have been used instead.
The arguments to IFDEFP are the same as with CONF. The first parameter is
the name of the field on which the decision is to be made. The second parameter
is the value that is to be tested against. The third parameter takes a value of
"1" or "0": 1 = print the line if the field contains the value; 0 = print the
line if the field does NOT contain the value.