Added a new bibconver formatting function "JOINMULTILINES".  It is used to join a multiline string into a single line.  The function interface is:    JOINMULTILINES(prefix,suffix) For each segment resulting from the split on newline, "prefix" will be appended to the front and "suffix" will be appended to the end. Example of calling:   For the field XX with the value:    Test    Case, A   And the function call:    <:XX^::XX::JOINMULTILINES(<subfield code="a">,</subfield>):>   The results would be:    <subfield code="a">Test</subfield><subfield code="a">Case, A</subfield>
Also added a new notation "^" to source fields, e.g.:
   <:XX^::XX:>
It is similar to the notation "*" which is used to signify repeating source
fields.  However, "^" is used to tell bibconvert to preserve the newlines
in the source field so that they can be processed by functions such as
"JOINMULTILINES", above.