Deprecated the following 3 functions:   + getInfo   + getInPending   + getInAlice
They basically search for details of an item that has at some point been
submitted for referee-approval. getInPending looks for those details in the
WebSubmit pending dir, whereas getInAlice looks for the details in the CDS
Invenio repository.  getInfo was some glue that first called getInPending,
then getInAlice if that failed to find anything.

The way they were used was like this:

try:
    (authors,title,sysno,newrn) = getInfo(doctype,categ,RN)
except TypeError:
    return _("Unable to display document.")

It's ugly because the function(s) either returns a 4-element tuple or integer
0, catching a type-error to realise that it found nothing.

Another problem (the major motivation for these changes) was that "getInPending"
searches in the pending directory for the following hard-coded filenames:
  + TI
  + TIF
  + AU
  + SN
This basically means that the Invenio admin isn't free to call her form fields
for title, etc as she likes, which isn't very nice. To get around this, I added
3 new functions:
  + get_pending_item_details (replaces getInfo)
  + get_brief_doc_details_from_pending (replaces getInPending)
  + get_brief_doc_details_from_repository (replaces getInAlice)

The idea is that in the "submit" stage of the item, a WebSubmit function
"Make_Dummy_MARC_XML_Record" should be called. This will create a MARC XML
record containing the metadata. This record can then be parsed here, and
fields like title, authors, etc extracted from it.

For backward compatibility reasons, "get_pending_item_details" fails to
find anything, the old "getInfo" is called. This should probably be removed
though.