BibSched graceful stop and new uniformize_bibupload_priorities.
New BibSched graceful directive to stop completely bibsched and all the running tasks.
Stopped tasks are now rescheduled (when reschedulable).
Fixed small bugs here and there.

If a biupload with priority 10 happens to be submitted
after a bibupload with priority less than 10 (say 9), the first
bibupload is blocked until the second is executed. To give real
priority to the first bibupload, we should rise the priority of the
second one up to the priority of the first.