Scheduler to wait for real start of tasks.
When a task is launched by bibsched, its status is set to SCHEDULED,
the task is executed in background and bibsched goes on with its
algorithm. Without waiting for the task to start real (e.g. by changing
its status, allocating a pid etc), race conditions might happen.
This patch (should) fix this problem.