Support for bibsched tasks with negative priority.
Previously tasks with priority less than 0 were never executed.
This was too strong, since it may be nice to give a task a priority like
"run when no other normal tasks are running". For this reasons now, if
you give a task a priority less than 0 (but higher or equal than -10),
it will be run after normal tasks.