New BibEdit in AJAX.
This is the first public release of a completely redone version of
BibEdit. This version uses AJAX technology to provide users with a much faster,
simpler and more responsive interface then the old version. It uses jQuery as
Javascript framework and the JSON protocol for data transfer.

The application is now served by the standard Invenio page handler and is
accessible through the URL /record/edit/ or optionally one can specify the
initial record to edit by going to /record/<recid>/edit. The old version is
kept as a backup an can be reached at /admin/bibedit/bibeditadmin.py.