WebSubmit: Send_APP_Mail function glitch fixed
In the Send_APP_Mail function, if the comments_file parameter
value IS set (in the websubmit gui), and a comment is NOT supplied by
the user for this specific submission (he is allowed to do so), an
exception is thrown, although the field might be non mandatory.
This commit fixes the problem.  Thanks to Theodoropoulos Theodoros.