The getter used some regular expression based on the string of the 'verb' parameter. The 'verb' in question here is 'ListRecords' and it also starts the initiation of a data block. However, xml allows explicit name spaces and thus this query fails in case we get qualified xml.
According to the specs of OAI-PMH the server should return a 'noRecordsMatch'
error. (Effectively, being the reason for the missing ListRecords stanza.). Thus
check for the existence of this error instead gives a valid check.

Additionally, the "parsing" fo the resumptionToken also assumed that
we do never get any namespace. This is hotfixed by a more general
regexp, that allows for the inclusion of namespaces, but ultimately
one should consider real XML parsing. (cf. FIXME)

Introduced to cope with OpenAIRE project set.