installation: add more fields to escape
I checked how say the password 'pass\'\'word' was rendering and if it could be escaped.
Python is automatically escaping this and actually creating this, pass'"'"''"'"'word.
Given that developers will see that this password is clearly not what they intended from the log message, this is probably ok, they can just change it.
Passwords such as 'pass''word' are rendered as 'password' due to python and it's concatenation.

So, for me, I think this function does what it is supposed to do. The small fix at least removes any ambiguity for developers.